[BROKEN] Move non-template configs to `raw` dir
NOTE: This commit makes it so that files in `raw` dir aren't copied to their corresponding locations.
This is due to me currently working on a rewrite of the installer in Rust, because Bash is making me go insane. I guess, "skill issue" :).

Consider not checking-out this until a corresponding [FIX] commit.